Wrote methods to find query nodes containing specific relations.
If it's a single relation, the first found by depth first search node is used.

If it's a list of relations, it's the first found by depth first search that has is itself in the case of a single relation in the list, or its children combine to contain all of the relations.